feat(oracle): split chatgpt client into dispatch/harvest primitives
